Refactor: consolidate all commands (#66)
