+++ src/overview/index.ts
@@ -3,12 +3,10 @@ import * as vscode from 'vscode';
 import { readFile as fsReadFile } from 'fs';
 import * as util from 'util';
 import * as path from 'path';
-const openurl: any = require('openurl');
 
 const readFile = util.promisify(fsReadFile);
 
 import { instrumentOperation, sendInfo } from "vscode-extension-telemetry-wrapper";
-import { validateAndRecommendExtension } from '../recommendation';
 
 let overviewView: vscode.WebviewPanel | undefined;
 const KEY_SHOW_WHEN_USING_JAVA = 'showWhenUsingJava';
@@ -68,34 +66,6 @@ export async function overviewCmdHandler(context: vscode.ExtensionContext, opera
       toggleOverviewVisibilityOperation(context, e.visibility);
     }
   });
-}
-
-export async function createMavenProjectCmdHanlder(context: vscode.ExtensionContext) {
-  if (!await validateAndRecommendExtension('vscjava.vscode-maven', 'Maven extension is recommended to help create Java projects and work with custom goals.')) {
-    return;
-  }
-
-  await vscode.commands.executeCommand('maven.archetype.generate');
-}
-
-// TODO: add entry to create standalone Java file
-
-export async function createSpringBootProjectCmdHandler(context: vscode.ExtensionContext) {
-  if (!await validateAndRecommendExtension('vscjava.vscode-spring-initializr', 'Spring Initializr extension is recommended to help create Spring Boot projects and manage dependencies.')) {
-    return;
-  }
-
-  await vscode.commands.executeCommand('spring.initializr.maven-project');
-}
-
-export async function showExtensionCmdHandler(context: vscode.ExtensionContext, operationId: string, extensionName: string) {
-  openurl.open(vscode.Uri.parse(`vscode:extension/${extensionName}`).toString());
-  // TODO: uncommment the below line after vscode address https://github.com/Microsoft/vscode/issues/62629
-  //vscode.commands.executeCommand('vscode.open', vscode.Uri.parse(`vscode:extension/${extensionName}`));
-}
-
-export async function openUrlCmdHandler(context: vscode.ExtensionContext, operationId: string, url: string) {
-  vscode.commands.executeCommand('vscode.open', vscode.Uri.parse(url));
 }
 
 export async function showOverviewPageOnActivation(context: vscode.ExtensionContext) {
