+++ src/utils/command.ts
